Miscellaneous build and compatibility fixes
- Use copy_c_str for the rangeproof order id in pro_backend, matching the other
  payment providers.
- Give the blind_version_sign C wrappers a uint64_t timestamp to match the
  header and C++ API.
- Remove the deprecated opt::quic_disable_mtu_discovery.
- Pass nullptr (not std::nullopt) for the _resync_clock callback argument.
- Locally suppress deprecation warnings around deprecated-path tests.
- Spell out value_or(core::callbacks{}) in the Core constructor so template
  argument deduction succeeds.
- Disable the Intel macOS CI builds.

+++ src/blinding.cpp
@@ -494,7 +494,7 @@ LIBSESSION_C_API bool session_blind25_sign(
 
 LIBSESSION_C_API bool session_blind_version_sign_request(
         const unsigned char* ed25519_seckey,
-        size_t timestamp,
+        uint64_t timestamp,
         const char* method,
         const char* path,
         const unsigned char* body,
@@ -520,7 +520,7 @@ LIBSESSION_C_API bool session_blind_version_sign_request(
 LIBSESSION_C_API bool session_blind_version_sign(
         const unsigned char* ed25519_seckey,
         CLIENT_PLATFORM platform,
-        size_t timestamp,
+        uint64_t timestamp,
         unsigned char* blinded_sig_out) {
     try {
         auto sig = session::blind_version_sign(
